Update dir structure, docs
