+++ web-console/src/lib/components/connectors/dialogs/KafkaInputConnector.tsx
@@ -5,7 +5,8 @@ import TabKafkaInputDetails from '$lib/components/connectors/dialogs/tabs/TabKaf
 import TabKafkaNameAndDesc from '$lib/components/connectors/dialogs/tabs/TabKafkaNameAndDesc'
 import TabLabel from '$lib/components/connectors/dialogs/tabs/TabLabel'
 import { ConnectorFormNewRequest, ConnectorFormUpdateRequest } from '$lib/services/connectors/dialogs/SubmitHandler'
-import { ConnectorType, connectorTypeToConfig, connectorTypeToIcon, parseKafkaInputSchema } from '$lib/types/connectors'
+import { connectorTypeToConfig, connectorTypeToIcon, parseKafkaInputSchema } from '$lib/functions/connectors'
+import { ConnectorType } from '$lib/types/connectors'
 import ConnectorDialogProps from '$lib/types/connectors/ConnectorDialogProps'
 import { ConnectorDescr, ConnectorId, NewConnectorRequest, UpdateConnectorRequest } from '$lib/types/manager'
 import { useEffect, useState } from 'react'
