+++ web-console/src/lib/components/streaming/builder/NodeTypes/OutputNode.tsx
@@ -2,7 +2,7 @@
 // program.
 
 import useNodeDelete from '$lib/compositions/streaming/builder/useNodeDelete'
-import { connectorDescrToType, connectorTypeToIcon } from '$lib/types/connectors'
+import { connectorDescrToType, connectorTypeToIcon } from '$lib/functions/connectors'
 import { Connection, NodeProps, Position, useReactFlow } from 'reactflow'
 
 import { Icon } from '@iconify/react'
