+++ web-console/src/lib/compositions/streaming/inspection/useTableUpdater.ts
@@ -4,7 +4,7 @@
 // In dbsp-speak this maintains an integral for a part of a relation.
 
 import { readLineFromStream } from '$lib/functions/common/stream'
-import { csvLineToRow, Row } from '$lib/types/ddl'
+import { csvLineToRow, Row } from '$lib/functions/ddl'
 import { NeighborhoodQuery, Relation } from '$lib/types/manager'
 import { parse } from 'csv-parse'
 import { Dispatch, SetStateAction, useCallback, useMemo } from 'react'
