+++ web-console/src/pages/_app.tsx
@@ -6,7 +6,7 @@ import StatusSnackBar from '$lib/components/common/errors/StatusSnackBar'
 import StandardVerticalLayout from '$lib/components/layouts/StandardVerticalLayout'
 import { usePageHeader } from '$lib/compositions/global/pageHeader'
 import themeConfig from '$lib/functions/configs/themeConfig'
-import { defaultQueryFn } from '$lib/types/defaultQueryFn'
+import { defaultQueryFn } from '$lib/services/defaultQueryFn'
 import { OpenAPI } from '$lib/types/manager'
 import Head from 'next/head'
 import { Router } from 'next/router'
